+++ src/utils/index.ts
@@ -26,4 +26,4 @@ export * from "./DepositUtils";
 export * from "./ValidatorUtils";
 export * from "./AddressUtils";
 export * from "./SpokeUtils";
-export * from "./MigrationUtils";
+export * from "./V3Utils";
